requirements.txt
python-novaclient>=2.23.0
python-ironicclient>=0.5.1